// frontend/src/api-client/model/reservation-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SpotReservation.Api | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const ReservationStatus = {
    Pending: 'Pending',
    Approved: 'Approved',
    Cancelled: 'Cancelled',
    Rejected: 'Rejected',
} as const;

export type ReservationStatus = typeof ReservationStatus[keyof typeof ReservationStatus];